requirements/common.txt
Django
lxml
mysql-python
hurry.filesize
django-haystack
pysolr
wsgiref
mock
django-autocomplete-light
sorl-thumbnail 
pillow
bitly_api
psycopg2
xlwt
unidecode
django-password-reset
openpyxl>=1.9
django-allauth
django-simple-captcha
django-haystack
